Provide DNS Search List in DHCH6_Reply
In case of an IPv6 only VM we have to serve a Domain Search List
so that Windows VM can find the kms server by issuing:

nslookup -type=srv _vlmcs._tcp

An extra config option in ipv6 section is added; domains
This is a list of search domains included in DHCP6_Reply.
Since we validate it with force_list() there is no need for a
trailing comma.

This info could be passed in RAs but DNSSL options is not currently
supported by scapy's inet6.py.

Signed-off-by: Dimitris Aragiorgis <dimara@grnet.gr>